Still no results

+++ TunerBass/Graph.h
@@ -41,7 +41,7 @@ public:
 	void clear();				//paint it black
 	void draw_axis();			//draw coordinate plane					
 	void draw_new_point(float x, float y);		//add new point to the graph
-	void draw_sample(float* sample, int offset);	//output sample to the graph
+	void draw_sample(int16_t* sample, int offset);	//output sample to the graph
 
 
 	void draw_text(int x, int y, char* text, int len);	//draws text on the screen using WiApi
